# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QCET E-Office Canonical API Contract
  version: 1.0.0
  description: >
    Canonical OpenAPI 3.1 contract for QCET E-Office (Trường Cao đẳng Kinh tế và Công nghệ Quảng Ninh).
    Documents public, authenticated, and privileged endpoints with strict authorization rules,
    standard error contracts, and DTO schemas.
servers:
  - url: http://localhost:3000
    description: Local development / test server
  - url: https://eoffice.qcet.edu.vn
    description: Institutional on-premise production server
security:
  - sessionCookie: []
paths:
...
  /api/documents:
    get:
      summary: List incoming and outgoing official documents
      operationId: listDocuments
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [INCOMING, OUTGOING]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            maximum: 100
            default: 20
      responses:
        '200':
          description: Document list
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Register incoming/outgoing official document (Clerk/Admin/Manager)
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentInput'
      responses:
        '201':
          description: Document registered
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: qcet_session
  schemas:
...
    CreateDocumentInput:
      type: object
      required: [title, documentNumber]
      properties:
        title:
          type: string
          maxLength: 300
        documentNumber:
          type: string
          maxLength: 100
        documentType:
          type: string
          enum: [INCOMING, OUTGOING]
        issuingAuthority:
          type: string
...
    ApiErrorBody:
      type: object
      required: [code, message, requestId]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: FORBIDDEN
        message:
          type: string
          example: Access denied to requested resource
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        requestId:
          type: string
          example: req-12345-67890
  responses:
...
    UnauthorizedError:
      description: Authentication required or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
    ForbiddenError:
      description: Authenticated but unauthorized for this object/action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorBody'
...